Optimized NavigatorAxis composition.
